Don't return a result when throwing
When there is no available activity for a launched
ActivityResultLauncher an ActivityNotFoundException is thrown, but the
underlying startActivityForResult APIs also still return a result with a
RESULT_CANCELLED request code.

We should ensure that when the ActivityResultRegistry's launch throws an
ActivityNotFoundException, we don't also return a result to the
callback.

RelNote: "The `ActivityResultRegistry` will no longer return a result
to the `ActivityResultCallback` when the `launch` call throws any
`Exception`."
Test: added tests
Bug: 200845664

Change-Id: If4f9159bce2edf7f3f0c76e7ff37ce21b05c13ee