Determine the PREVIEW size by the max display from DisplayManager#getDisplays()
Originally, the PREVIEW size was determined by the default display retrieved by WindowManager#getDefaultDisplay(). It will cause IllegalAccessException when StrictMode is enabled because of accessing visual service WindowManager from a non-visual context. The exception won't happen when using DisplayManager to get the max display and then obtain the display real size from it.

Bug: 201002367
Test: ./gradlew bOS
Change-Id: I7f9eaaf85fbfe053bceb3b4cd58596d67e560982