Rename WindowLayoutInfo to TestWindowLayoutInfo
Rename the WindowLayoutInfo function in the testing artifact to
TestWindowLayoutInfo. Changing the name will eliminate the possible
conflict in the core library.
Make the display feature list optional in
WindowLayoutInfoTesting#createWindowLayoutInfo by adding a default value of an
empty list.
Add tests for the TestWindowLayoutInfo function.
Add Java overloads so Java developers do not have to call with null
values.

Relnote: Change name to TestWindowLayoutInfo.
Bug: 201382530
Test: ./gradlew window:window-testing:cAT
Change-Id: I4a2fd3a8da2e40fbabcb72c4f323f0090a573bd4