Let apps disable MR2 dynamically
Add a flag to MediaRouterParams to disable MR2 from the app.
It can disable MR2 feature even when MediaTransferReceiver
is declared.
It is not supposed to be changed at runtime dynamically
but it was okay during a manual test.

Relnote: "Add a flag into MediaRouterParams that can be used to disable
seamless transfer at runtime"

Bug: 195331360
Test: ./gradlew mediarouter:mediarouter:connectedCheck &&
run support v7 demos manually, by using
OutputSwitcherActivity and LegacyMediaRouterActivity repeatedly

Change-Id: I53d687a07bceb307530241851bb453d425f20ba1