Add text alignment parameter.
I copied the interface used by Jetpack compose for this.
Sadly, justified text will only be available on Android S+ for App
Widgets. To allow changes in locale to be reflected, locale change
broadcast are also interpreted correctly by the GlanceAppWidgetReceiver.

On this CL, the following classes where moved in their own file and
package, to reduce the size of the file linked to the Text composable:

- FontStyle
- FontWeight
- TextAlign
- TextDecoration
- TextStyle

Relnote: n/a
Test: ./gradlew :gradle:gradle-appwidget:test
Test: Integration tests added
Bug: 200024433

Change-Id: I419fbb88c10f9c519a2879a94870fcc1406a28a2