InlineSlider implementation for wear compose
An InlineSlider is a control which can increase or decrease some value.
User sets a float range (min/max values) and a number of segments.
Step value is calculated as a difference between min and max, divided by a number of segments.
Each click on left or right buttons then changes value -step and +step accordingly.

Customization like background color and a color of a bar is also available.
This slider supports a visible segmentation.

Bug: 200919272
Test: Run tests in androidx.wear.compose.integration.demos.test
Relnote: "Added InlineSlider for wear compose. InlineSlider allows users make a selection from a range of values. The range of selections is shown as a bar between the minimum and maximum values of the range, from which users may select a single value."
Change-Id: If01485d8dd34fc309770c843709a8bceae10e2fd