Add transform for `SuspendFunction`s as parameters
Currently, `KSType.typeName` doesn't consider suspend functional types
a special case. This means that `suspend () -> Unit` would get resolved
as `Function1<Unit>` and not as `Function1<Continuation<Unit>, ?>`.
This CL adds a check if the `KSType` is a suspend functional type
and adds a continuation parameter with the correct return type.

Test: MethodSpecHelperTest#kotlinParametersAsFunction, SuspendingQueryTest
Bug: b/201674894
Change-Id: I02c97b4ee87888c1494605d8fce2442e9e0ce017