Add test to verify mapping between new tables and migrated ones.
This test verifies that Room mapping code works with both newly created tables and altered tables whose star projections results in a different order of columns.

For example, start with entity with fields: A, B, C. Then, on the entity, a new field and column is added with the fields order being A, B, X, C along with a migration being 'ALTER TABLE _ ADD COLUMN X _'. The column result order for a star projection query will be different between a migrated database and a newly created one, but Room should be able to do the right mapping anyway.

Test: AlteredTableColumnOrderTest
Change-Id: I938b4373642d48658181e2df099b7d7fde2d4fad