Deprecate Typeface backed by FontFamily
This concept is unsupportable with async fonts, and will be replaced
with a new API for font preloading.

- Deprecate Typeface(FontFamily) constructor as it internally loads
fonts syncronously on call. This doesn't fit with async font loading
and fallback chains.
- Deprecate fontResource(FontFamily): Typeface, as it doesn't allow for
a fallback model in the presence of downloadable fonts and may block for
10+ seconds.

Test: n/a (just a deprecation)
Relnote: "`Typeface(FontFamily)` constructor is deprecated. This was
previously used to preload fonts, which may take up to 10 seconds for
downloadable fonts. With downloadable fonts, this call may block for 10
seconds. Instead use FontFamilyResolver.preload"
Relnote: "`fontResource(FontFamily): Typeface` is deprecated. This was
previously used to preload fonts, which may take up to 10 seconds for
downloadable fonts. Instead use FontFamilyResolver.preload"
Bug: b/174162090

Change-Id: If8e7c6ce7cd64be8094a576587cc1329e19d246f