Fix fail to filter by Namespace in remove.
Framework-appsearch in Android S will disable the given namespace filter
if it is not empty and none of given namespaces exist.
And that will result in Icing remove documents under all namespaces if
it matches query express and schema filter.If none of user given
namespace filter exist, remove will remove all.

Bug: 197361770
Test: AppSearchSessionTest
Test: manually tested on Android S device.
Change-Id: If9228876f6d051935a03daa09c8e7074182b39f0