Add require not null support for exact type conversions
This CL updates NullAwareTypeConverter to have the ability to inject
RequireNotNull checks when converting between two explicit types.

Whether we want to add this or not is debateable but we do it when
reading from Cursor so this does make it more consistent.

For the added tests, the cursor test already passes without the fix
whereas the `String!` to `MyClass!` exact conversion cases fail. Hence
making that also pass makes the converter store more consistent.

The downside is that, developer might miss a real case and receive NPE
at runtime.

Bug: 203422633
Test: NullabilityAwareTypeConverterStoreTest
Change-Id: Idb625c75f9c0bbbf1040737dcdc3116e616fe6cf