Robust handling for disappearing pointers
Bug: 202410799
Bug: 201556405

awaitDragOrUp() assumes that all pointers that have gone
down will send an "up" event and not just disappear.

When a new device replaces an existing device, a
"cancel" is now sent to ensure that an "up" has
been received. The cancel event now doesn't send
any event if there were no "pressed" pointers and
sends all pointers if there was a "pressed" pointer.

This is a blind, but safe fix. I cannot reproduce the
original problem, but it appears as if events from
a different input are not causing the ACTION_CANCEL
for a previous event.

Test: new test, modified test
Change-Id: I8b2e22d3700be17750ede612b9cafd229627fa92