Fix for crash when layoutNodeWrappers are reused
When ModifierLocalConsumerNodes are reused, if they were
generated by a modifier that implemented multiple interfaces,
the onModifierLocalsChanged() for the other LayoutNodeWrappers
could be called before the LayoutNodeWrapper is initialized.
This CL makes sure that the ModifierLocalConsumerNode wraps all
the other LayoutNodeWrappers created along with it (By a modifier
that implements multiple interfaces).

It also fixes a bug where the Focus and KeyInput LayoutNodeWrappers
were not initialized properly when the LayoutNodeWrapper was resused.

Bug: 203280574
Bug: 203198844
Fixes: 203280574
Fixes: 203198844
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.node.LayoutNodeWrapperInitializationTest
Change-Id: I99a402f00866fa05290cdc6db1df8b830e4333d2
(cherry picked from commit c971426797e38b7302cea047d5c15a86d395d766)