Add ComponentDialog
Provide a stripped down equivalent to
ComponentActivity for a Dialog that implements
LifecycleOwner and OnBackPressedDispatcherOwner.

Importantly, this connects the dialog's
OnBackPressedDispatcher to the Dialog
onBackPressed() API, thus ensuring that it
is called while the dialog is on the screen.

Relnote: "Added `ComponentDialog`, a subclass of
`Dialog` that includes an `OnBackPressedDispatcher`
that will be called when the system back button is
pressed when the dialog is visible. Importantly,
this subclass also sets the
`ViewTreeOnBackPressedDispatcherOwner`, allowing views
a generic way to retrieve the correct dispatcher
whether it exists in a `ComponentActivity` or a
`ComponentDialog`."
Test: ComponentDialogTest suite
BUG: 205992875

Change-Id: I8a1bcd3c4e2b4544d4b4460e2e7fd7424af34452