Resolving issue in relational query methods occurring specifically in LEFT JOIN queries.
The issue here arises when the query result adapters create empty value objects for the key-value pairs when the value is not found in the cursor. This results in invalid entries in maps in both 1-1 and 1-many relations, where a valid key is mapped to basically an object where all of its fields are null. This fix implements an additional column null check on the value object's columns and avoids adding the entry to the map if any null columns are found.

In the case where a 1-many mapping is present, the collection returned for a key will not include the invalid value object if it is not found in the cursor. If no valid values are found, then a key will be mapped to an empty collection.

Test: Added test cases for a 1-1 and 1-many mapping of a left join MultimapQueryTest.java.
Bug: 201946438
Change-Id: Id5552b9a7766e43f9d0434d7e5d6aff9dc84afb5