Add API to force-enable app tag tracing
Fixes: 182386956
Test: TrivialStartupFullyDrawnBenchmark # on user build, fully drawn metric missing without forcing the tag
Relnote: """Add Trace.forceEnableAppTracing() an API to force-enable
app trace section capture (i.e. android.os.Trace / andoridx.tracing
APIs) on non-debuggable builds. Call this at the beginning of startup
to enable non-debuggable accurate system tracing prior to API 29."""

This is done as a single "forceEnable" call instead of a setter as
there's no way with reflection to read the current value, and a setter
allows you to clear the value set in the manifest, which makes
ApplicationInfo#isProfileableByShell less useful for
e.g. macrobenchmark.

Change-Id: I3a30949580d8db6777302a759b041f9480e04e72