Correctly open pre-package database during PrepackagedDatabaseCallback invocation.
The actual pre-package database was not being opened because only the file name was being used when creating the temporary open helper instead of the full path. The framework open helper will only open an existing file if the name of the database is a path (starts with '/') otherwise it create a new file. This changes fixes it by using the absolute file path as name.

This change also fixes the scenario where the pre-package database has a zero version. The framework open helper does not allow this so to workaround Room will open it as if the version was 1 and then reset it back to 0 once the DB is opened and handed to the callback.

Bug: 203885584
Test: PrepackageTest
Change-Id: I1bda487de3f335a3346524fa7be9d51ff0daac6a