Add ScalingLazyItemScope and fillParentMaxSize/fillParentMaxWidth/fillParentMaxHeight modifiers.
Adds functionality already available in the LazyRow/Column but missing from ScalingLazyColumn.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 205276473
Relnote: "Added a ScalingLazyItemScope and some new modifiers fillParentMaxSize/fillParentMaxWidth/fillParentMaxHeight to allow list items to be sized based on the size of the parent container. Items can be set to fill all or a fraction of the parents size. This exposes functionality already available in the LazyRow/Column"
Change-Id: I4612f5088090567fe9fcec598ffbd04642fa9aa5