Creates RearDisplay Mode API and the WindowAreaController
Enables the RearDisplay Mode functionality to move the current window
to the display that is aligned with the rear camera. Currently the only
public API is the API exposed in the window-java package as we continue
to iterate on the kotlin API we want to expose.

Bug: 207686851
Test: WindowAreaControllerImplTest, RearDisplayStatusUnitTest
Relnote: "Introduces the WindowAreaController and API's to enable RearDisplay Mode to
move the current window to the display that is aligned with the rear camera."

Change-Id: I388ab775740713650e6142135fbfc57891d2214b