Attach ComponentActivity's SavedStateRegistry in init
By using SavedStateRegistryController's new
performAttach() method, we can ensure that the
LifecycleObserver added by that method is added
to the ComponentActivity as part of its initialization.

This ensures that its LifecycleObserver runs before
any LifecycleObserver added by other initializers
(such as those from registerForActivityResult() or
those manually added in an init block).

Relnote: "Fixed a crash when accessing a `ViewModel`
for the very first time from a
`registerForActivityResult()` callback or the
callbacks to a `LifecycleObserver` added as part of
`init` of a `ComponentActivity`."
Test: savedStateEarlyRegisterInitAddedLifecycleObserver
BUG: 177246258

Change-Id: Ife83f26bedddb2b843a2ce977d0e9526f070ac07