Allow nullable boolean array in safe args
While you should not be able to give a boolean a default value of null
you should be able to make an array default to null. We need to enable
safe args to generate boolean array with nullable values.

RelNote: "Fixed a bug in safe args that would not allow boolean arrays
to have a `null` value."
Test: modified NavParserTest
Bug: 174787525

Change-Id: I8c396c8f174d5924bb5ee147261de158256fba94