Fix crash in MeasureAndLayoutDelegate.forceMeasureTheSubtree()
It was happening because we were asserting that the node will not be in NeedsRemeasure state anymore as a result. However it is not always true as we can skip remeasuring for nodes which are not placed and can't affect the parent size. The updated logic will correctly skip such unplaced subtrees.

Fixes: 206732855
Test: new test in NodesRemeasuredOnceTest
Change-Id: I37abed763d269b51b858a86c820fd64038eca505