Add method to programatically select an option on the PickerState
This allows developers to pick the option they need, useful for example
for more complex compound pickers like a date picker.
It's also possible to set the initial option to be selected when creating
the state.

Relnote: "Add method to programatically select an option on the PickerState,
the initially selected option can now also be specified when creating a
PickerState"
Test: Added.
Bug: http://b/199476864

Change-Id: I92bdf1ba853a6f53ff1264dbdbe6b4d0f6128941