Introduce a complications cache
Test: Manual testing
Bug: 196958807
Relnote: We've introduced a per-watchface ComplicationData cache the purpose of which to allow the watch face to display last known complication data values upon loading until the system has had a chance to update them. Complications form an integral part of some watchfaces and they look very spartan without them. There is a new API method WatchFaceControlClient.hasComplicationCache. Intended for OEMs this may influence the system's strategy for sending complications to a watchface. In addition ComplicationData has an isCached property and it is recommended that cached complications are rendered differently because the tapAction can not be cached and will be null in a cached complication.
Change-Id: I404b06e624e745b0ae5d48558658825c9c1cc7f3