Fix LazyListState.animateScrollToItem() skipping the item on slow devices
It was possible that we go into a wrong if branch in the scrolling logic when the overshoot happened and the list ended up scrolled to the end instead of pointing to the needed element.
Unfortunately I don't know how to cover it with a test, we don't have such a granular control over the scroll animation from the tests.

Test: manually, with the sample provided by the user reported the bug
Fixes: 200577808
Change-Id: Ie55adbd9d360c46362e96d125777e70e5c4334c6