Request layout when tracking AnimatedVisibility transitions
Some animation types, e.g. slide, require a measure/layout pass to be
set up correctly. We noticed that, sometimes, slide animations were not
being detected by Animation Preview in Android Studio. That was caused
by a timing issue when tracking AnimatedVisibility transitions. If the
ComposeViewAdapter measure pass happened on a frame after we read the
animations, we would lose track of them. This is now fixed by calling
ComposeViewAdapter#requestLayout when tracking the animations, which
guarantees that measure/layout is executed after the animations are
created and before we read them for the first time.

Fixes: 206461065
Test: PreviewAnimationClockTest + test on Android Studio side
Change-Id: I89c6c68f53aa0b9b7e26d4ab90abcdc8811087e9