Improve self referencing type handling
This CL fixes two issues on how we handle self referencing types.

First one happens when generating type names for parameter / property
types.
The previous fix for TypeVariableNames do not cover this case because
the TypeArgument always refereences another type that still has type
arguments.
This one is fixed by simply using <?> for star projection. It was
previously not possible to use * because we always defaaulted to star
projection when resolving types. We don't do it anymore since the last
variance resolution fix so we can now simply use ? when we encounter a
star variance.

The second one happens in variance resoltuion where we would infinitely
try to inherit variance for method parameters that have self referencing
types. A similar "previoulsy seen types" solution is also added there to
avoid this issue.

There is also a small improvement for raw types in Java sources. KSP
always resolves raw types to their star projection (e.g. List becomes
List<Object>). I've added a KSType.isRaw check into KSTypeExt to detect
this so that we can better match JavaPoet behavior.
Note that, isRaw works by running toString on the type, which is not great
but appearantly, KSP itself does it too so why not ? :)

Bug: 208207043
Test: XTypeTest KspTypeNamesGoldenTest

Change-Id: Ia6657c5a0c8d08cdc92aab2dd9cd86d3e75d0bcb