Ensure Dialogs destination are properly restored
When there is a config change or process death while a dialog
destination is on top of the stack, NavHost is recomposed twice, once
after the initial restoration and again after the navController get to
the proper destination. During the first composition, NavHost calls
DialogHost and it has the entry associated with the dialog, but since
since the Lifecycle has not yet reached at least STARTED, the dialog is
not shown. On the second composition, the Lifecycle of the entry is
STARTED, but since we have already called the DialogHost function with
the dialogNavigator, we don't actually execute the function cause from
the perspective of compose, nothing has changed.

We should incorperate the concept of visible entries into DialogHost so
that navigation properly takes into account the changes in Lifecycle
states that determine the appropriate screen.

This also mean if you make a dialog your start destination, it will now
show properly.

RelNote: "Dialog destination are no properly restored on top of the
screen after config changes or process death."
Test: NavHostTest
Bug: 207386169

Change-Id: I4c0dce39bbbe26d05d0cb60c65c6e5fcfbbc62b2