Hook up removeByQuery to send change notifications to AppSearchObserver.
This is implemented by doing a regular query if there are observers, and
issuing delete-by-id calls for each returned document.

This could be improved later by having Icing Lib return information about
the URIs, namespaces and types of the documents that were affected. This
would require Icing Lib to store an additional map for each document.

Bug: 193494000
Test: testRegisterObserver_removeByQuery
Change-Id: I1293e071edd390effb37e822bf3c1336af384432