Invert PointerEvent.scrollDelta.x on Android
We should invert only by y.

If we do that, the behaviour would be the same, as in Web and Swing.

P.S. Web has two scrolling API's:

-wheelDelta, Deprecated https://developer.mozilla.org/en-US/docs/Web/API/Element/mousewheel_event
Behaviour as in Android (y is inverted)

-deltaY https://developer.mozilla.org/en-US/docs/Web/API/Element/wheel_event
Behaviour as in Swing (y isn't inverted)

Test: https://jsfiddle.net/27zwteog

When y isn't inverted, it looks more natural - when we scroll wheel up - we move the wheel touch point up.
And if we work in the usual coordinate system, the offset should be -1.

Test: manual
Relnote: PointerEvent.scrollDelta.y is now inverted on Android (now it returns 1 instead of -1 if we tilt mouse wheel to the right)
Change-Id: Ia981193e0870111d3f595b5b6c14f6bdc98f9873