Properly handle wildcards for sealed classes
This CL fixes a bug where we didn't use wildcards for sealed classes
because they are not `open` but in jvm, they actually are for their
subclasses.

Eventualy we'll replace this code with the new jvmWildcard API in KSP
but unfortunately it has certain bugs that blocks migration for now.
https://github.com/google/ksp/issues/773
https://github.com/google/ksp/issues/772

Bug: 210299317
Test: KspTypeNamesGoldenTest
Change-Id: I1a16b8c28bdf68eced1b766964e29a1128b0b2ef