Introduce Profile rules to Wear Compose modules.
See aosp/1728275 for information about ART profiles
and their use to improve application performance.

Test: N/A

Bug: 209956522

Relnote: "Added ART profiles for Wear Compose libraries.
ART can then leverage each profile when the APK is installed
on devices in order to ahead-of-time compile a specific
subset of the application to improve the performance of
the application, especially the first run.
Note that this will have no effect on debuggable applications."

Change-Id: I0fc31a010af5426fefbab2685efa467fb108d295