AndroidX Webkit: make IdlingResource NonNull & final
This is a speculative fix for the flakiness observed on the bots. The
bots are reporting that our UriIdlingResource gets into an invalid state
where its internal counter is negative. One possible explanation would
be if mUriIdlingResource is null before the load (so we don't increment)
but non-null after the load. Marking this field as final and
initializing it eagerly is a robust way to prevent this, so this change
shouldn't hurt and it simplifies the code.

Bug: 211597907
Test: ./gradlew :webkit:integration-tests:testapp:connectedCheck (1000 runs)
Change-Id: Ic54ee0de5023a80691e72e9669d8f5c505f4f357