Fix LoggerTest test failures
To avoid the LoggerTest being affected by unexpected log items, the test will only check the items of the specified tag. If a log item should be dumped, then, it can be found by the tag. If a log item should not be dumped, then, no log item with the specified tag should be found in the remaining log items.

Bug: 207674161
Test: LoggerTest
Change-Id: If4cd7c4664d34c55148111cf186a5a85ee2262c0