Additional integer API for InlineSlider and Stepper
One of issues with Slider & Stepper is that internally it uses a floating point math which leads to precision errors.
For example instead of 500 - component can return 499.999. If developers use .toInt() instead of roundToInt(), they will get 499 integer.
https://kotlinlang.slack.com/archives/C02GBABJUAF/p1638873187094400

Possible solution of this problem is to add an Integer api for such use cases

Bug: 209734903
Test: Added integer tests for InlineSlider and Stepper
Relnote: "Additional integer API for InlineSlider and Stepper"
Change-Id: I1b5d66881cec976f5d2c2c3fb272f9eabd5a8ddd