Add optional support for named ComplicationSlots
Re AutoBoxing, none of these code changes are in 'hot' paths so
the performance and memory implications of using boxed types
doesn't really matter.

Bug: 194771277
Test: Extended existing tests
Relnote: We've added setNameResourceId & setScreenReaderNameResourceId (which reference string resoures) to ComplicationSlot.Builder and corresponding getters in androidx.wear.watchface.client.ComplicationSlotState.  This allows the system to fetch the names of ComplicationSlots for use in editors and screen readers.
Change-Id: If6c6a13d5af8392269c11e19ecfc41542c195f6c